pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faster_app"
version = "0.0.39"
description = "一个轻量级的 Python Web 框架, 提供自动发现、模型基类、命令行工具等功能"
readme = "README.md"
license = "MIT"
authors = [{ name = "peizhenfei", email = "peizhenfei@hotmail.com" }]
maintainers = [{ name = "peizhenfei", email = "peizhenfei@hotmail.com" }]
keywords = ["web", "framework", "fastapi", "tortoise", "orm", "cli"]

requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "fastapi-pagination>=0.14.1",
    "fire>=0.7.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "rich>=14.1.0",
    "tortoise-orm[asyncpg]>=0.25.1",
    "uvicorn>=0.35.0",
]


[project.urls]
Homepage = "https://github.com/mautops/faster_app"
Documentation = "https://github.com/mautops/faster_app#readme"
Repository = "https://github.com/mautops/faster_app.git"
"Bug Tracker" = "https://github.com/mautops/faster_app/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["faster_app*"]
exclude = ["tests*", "docs*", "__pycache__*", "*.pyc", "*.pyo"]

[tool.setuptools.package-data]
"faster_app" = ["statics/*", ".env.example", "runtime/*"]

[project.scripts]
faster = "faster_app.cli:main"

[dependency-groups]
dev = [
    "build>=1.3.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.21",
    "mkdocstrings[python]>=0.30.1",
    "twine>=6.2.0",
]
